Improved saving system
I was thinking of making the settings save with the .pt3d file, now it does save the settings in the same file. If this was not implemented I would have to check in the user dir every runtime and delete the settings file of the.pt3d files deleted from the downloads by the user.